clean up.  I still define my own variables in a .pro file, they just never seem to take.
git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@1179 640d5228-2204-0410-8d71-8f46fa6850b4

+++ rtplot/extern.h
@@ -22,11 +22,11 @@ extern std::vector<RAWTBL *> raw;          /* Alphabeticly sorted pointers */
 
 extern DEFAULT *Defaults[];
 
-extern int	nVariables;
+extern size_t	nVariables;
 extern VARTBL	Variable[];
 
 
-extern int	NumberDataSets, NumberSeconds, NumberElements[], CurrentDataSet;
+extern size_t	NumberDataSets, NumberSeconds, NumberElements[], CurrentDataSet;
 extern int	Aircraft;
 
 extern char	*ADSrecord;
